requirements.txt
torch
torchvision
torchaudio
facenet-pytorch
opencv-python
ultralytics
pillow
numpy